removed the donteat page and replaced with update donate page
