+++ next.config.js
@@ -13,7 +13,6 @@ const pages = [
   '/contact',
   '/terms',
   '/privacy-policy',
-  '/donteat',
   '/hubs',
   '/home',
 ];
